requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
asyncpg
sqlalchemy[asyncio]
redis
python-jose[cryptography]
httpx
python-dotenv
alembic
pytest
psycopg2-binary
pytest-asyncio